adding choices files to each example
