+++ code/drasil-example/gamephysics/gamephysics.cabal
@@ -1,10 +1,10 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.34.2.
+-- This file has been generated from package.yaml by hpack version 0.33.0.
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 866ab6e1cd802a73f6b1f04d84d5a4cc211d4b80cec09167e96effa67d85c493
+-- hash: 6afb02eb1bedc2c0cf95057fe76d2d5717762421143645feefea481040693bf9
 
 name:           gamephysics
 version:        0.1.24.0
@@ -25,6 +25,7 @@ library
       Drasil.GamePhysics.Assumptions
       Drasil.GamePhysics.Body
       Drasil.GamePhysics.Changes
+      Drasil.GamePhysics.Choices
       Drasil.GamePhysics.Concepts
       Drasil.GamePhysics.DataDefs
       Drasil.GamePhysics.Expressions
