+++ code/drasil-example/glassbr/lib/Drasil/GlassBR/Choices.hs
@@ -1,0 +1,32 @@
+module Drasil.GlassBR.Choices where
+
+import Language.Drasil.Code (Choices(..), CodeSpec, codeSpec, Comments(..), 
+  Verbosity(..), ConstraintBehaviour(..), ImplementationType(..), Lang(..), 
+  Logging(..), Modularity(..), Structure(..), ConstantStructure(..), 
+  ConstantRepr(..), InputModule(..), AuxFile(..), Visibility(..),
+  defaultChoices)
+
+import Drasil.GlassBR.ModuleDefs (allMods)
+import Drasil.GlassBR.Body (fullSI)
+
+code :: CodeSpec
+code = codeSpec fullSI choices allMods
+
+choices :: Choices
+choices = defaultChoices {
+  lang = [Python, Cpp, CSharp, Java, Swift],
+  modularity = Modular Separated,
+  impType = Program,
+  logFile = "log.txt",
+  logging = [LogVar, LogFunc],
+  comments = [CommentFunc, CommentClass, CommentMod],
+  doxVerbosity = Quiet,
+  dates = Hide,
+  onSfwrConstraint = Exception,
+  onPhysConstraint = Exception,
+  inputStructure = Bundled,
+  constStructure = Inline,
+  constRepr = Const,
+  auxFiles = [SampleInput "../../datafiles/GlassBR/sampleInput.txt", ReadME] 
+}
+
