+++ code/drasil-example/pdcontroller/pdcontroller.cabal
@@ -1,10 +1,10 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.34.2.
+-- This file has been generated from package.yaml by hpack version 0.33.0.
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: d652777950cb4b0613874ad3eb06236f164f05cbfe87633a4484fd8046a80123
+-- hash: 08ccf53062f0756b2f945d94a5663c4e70470b19420d8796a33346037478d5c1
 
 name:           pdcontroller
 version:        0.1.24.0
@@ -25,6 +25,7 @@ library
       Drasil.PDController.Assumptions
       Drasil.PDController.Body
       Drasil.PDController.Changes
+      Drasil.PDController.Choices
       Drasil.PDController.Concepts
       Drasil.PDController.DataDefs
       Drasil.PDController.GenDefs
