+++ code/drasil-example/ssp/lib/Drasil/SSP/Choices.hs
@@ -1,0 +1,31 @@
+module Drasil.SSP.Choices where
+    
+-- import Language.Drasil (QDefinition)
+-- import Language.Drasil.Code (Choices(..), CodeSpec, codeSpec, Comments(..), 
+--   Verbosity(..), ConstraintBehaviour(..), ImplementationType(..), Lang(..), 
+--   Modularity(..), Structure(..), ConstantStructure(..), 
+--   ConstantRepr(..), InputModule(..), matchConcepts, AuxFile(..), 
+--   Visibility(..), defaultChoices)
+-- import Drasil.SSP.Body (fullSI)
+
+-- code :: CodeSpec
+-- code = codeSpec fullSI choices []
+
+-- choices :: Choices
+-- choices = defaultChoices {
+--   lang = [Python, Cpp, CSharp, Java],
+--   modularity = Modular Combined,
+--   impType = Program,
+--   logFile = "log.txt",
+--   logging = [],         -- LogVar, LogFunc
+--   comments = [],    -- CommentFunc, CommentClass, CommentMod
+--   doxVerbosity = Quiet, -- Verbose, Quiet
+--   dates = Hide,      -- Show, Hide
+--   onSfwrConstraint = Warning,  -- Warning, Exception
+--   onPhysConstraint = Warning,  -- Warning, Exception
+--   inputStructure = Unbundled,    -- Unbundled, Bundled
+--   constStructure = Inline,   -- Inline, WithInputs, Store Structure
+--   constRepr = Const,    -- Var, Const
+--   conceptMatch = matchConcepts ([] :: [QDefinition]) [],
+--   auxFiles = [SampleInput "../../datafiles/SSP/sampleInput.txt"]
+-- }
