+++ code/drasil-example/template/app/Main.hs
@@ -1,9 +1,7 @@
 module Main (main) where
 
 import GHC.IO.Encoding
-
 import Language.Drasil.Generate (gen, DocSpec(DocSpec), DocType(SRS, Website))
-
 import Drasil.Template.Body (srs, printSetting)
 
 main :: IO()
