+++ code/drasil-website/stack.yaml
@@ -44,6 +44,16 @@ packages:
 - ../drasil-printers
 - ../drasil-theory
 - ../drasil-utils
+- ../drasil-example/dblpendulum
+- ../drasil-example/gamephysics
+- ../drasil-example/glassbr
+- ../drasil-example/hghc
+- ../drasil-example/nopcm
+- ../drasil-example/pdcontroller
+- ../drasil-example/projectile
+- ../drasil-example/ssp
+- ../drasil-example/swhs
+- ../drasil-example/template
 # Dependency packages to be pulled from upstream that are not in the resolver.
 # These entries can reference officially published versions as well as
 # forks / in-progress versions pinned to a git hash. For example:
